Refactored to accomdate for regional differences.
We have decided to franchise the pizza store and needed to support regional differences in pizzas.

We've create a new factory for each regional type of pizza, but this approach had a few issues.

Franchies were utilizing the factories for creating the pizzas,
but thy began to employ their own homegrown procedures for the rest of the process.

I think what we need is a framework that ties the store and the pizza creation together, yet stil allows for things to remain flexible.