A new framework for the pizza store.
We put the createPizza() method back into the Store class but as an abstract method, and then created a Store subclass for each region.

We're pushing all the variation on how pizzas are made by region into the createPizza() method and make each store responsible for implementing the method themselves.