Abstracted prepareRecipe method.
We've recognized that the two recipes are essentially the same, although some of the steps require different implementations.

So we're generalized the recipe and placed it the base class.

CaffineBeverage knows and controls the steps of the recipe, and performs steps 1 and 3 itself, but relies on Coffee or Tea to do steps 2 and 4.